changes for pagination
* removed newly added pagination utils and added same logic inside already existed utils.py
* implemented pagination changes in all applicable methods
